+++ .github/workflows/main.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Setup uv
         uses: astral-sh/setup-uv@v7
         with:
-          version: "0.9.2"
+          version: "0.9.17"
       - name: Create virtual environment
         run: uv sync --only-dev
       - name: Publish the docs to GitHub Pages
