+++ .github/workflows/tag.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Setup uv
         uses: astral-sh/setup-uv@v7
         with:
-          version: "0.9.2"
+          version: "0.9.17"
       - name: Build source dist and wheels
         run: uv build
       - name: Upload source dist and wheels to artifacts
@@ -58,7 +58,7 @@ jobs:
       - name: Setup uv
         uses: astral-sh/setup-uv@v7
         with:
-          version: "0.9.2"
+          version: "0.9.17"
       - name: Publish source dist and wheels to PyPI
         run: uv publish
 
@@ -109,7 +109,7 @@ jobs:
       - name: Setup uv
         uses: astral-sh/setup-uv@v7
         with:
-          version: "0.9.2"
+          version: "0.9.17"
       - name: Create virtual environment
         run: uv sync --only-dev
       - name: Publish the docs to GitHub Pages
