+++ .pre-commit-config.yaml
@@ -15,20 +15,20 @@ repos:
       - id: check-added-large-files
       - id: check-merge-conflict
   - repo: https://github.com/crate-ci/typos
-    rev: "v1.38.1"
+    rev: "v1.40.0"
     hooks:
       - id: typos
   - repo: https://github.com/astral-sh/uv-pre-commit
-    rev: "0.9.2"
+    rev: "0.9.17"
     hooks:
       - id: uv-lock
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.14.0"
+    rev: "v0.14.9"
     hooks:
       - id: ruff-check
       - id: ruff-format
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: "v1.18.2"
+    rev: "v1.19.0"
     hooks:
       - id: mypy
         additional_dependencies:
