+++ changelog.d/13.changed.md
@@ -1,0 +1,1 @@
+Replace supplementary record/manager classes with mixins
