+++ docs/managers/index.md
@@ -873,9 +873,7 @@ The managers for these record types have additional methods for querying records
 * [Currencies](currency.md)
 * [OpenStack Customer Groups](customer-group.md)
 * [OpenStack Grant Types](grant-type.md)
-* [Partner Categories](partner-category.md)
 * [Pricelists](pricelist.md)
-* [Product Categories](product-category.md)
 * [OpenStack Reseller Tiers](reseller-tier.md)
 * [Sale Orders](sale-order.md)
 * [OpenStack Support Subscription Types](support-subscription-type.md)
