+++ openstack_odooclient/managers/reseller_tier.py
@@ -17,11 +17,16 @@ from __future__ import annotations
 
 from typing import Annotated
 
-from ..base.record import ModelRef, RecordBase
-from ..base.record_manager_named import NamedRecordManagerBase
+from ..base.record.base import RecordBase
+from ..base.record.types import ModelRef
+from ..base.record_manager.base import RecordManagerBase
+from ..mixins.named_record import NamedRecordManagerMixin, NamedRecordMixin
 
 
-class ResellerTier(RecordBase["ResellerTierManager"]):
+class ResellerTier(
+    RecordBase["ResellerTierManager"],
+    NamedRecordMixin["ResellerTierManager"],
+):
     discount_percent: float
     """The maximum discount percentage for this reseller tier (0-100)."""
 
@@ -82,14 +87,14 @@ class ResellerTier(RecordBase["ResellerTierManager"]):
     under this tier.
     """
 
-    name: str
-    """Reseller tier name."""
-
     min_usage_threshold: float
     """The minimum required usage amount for the reseller tier."""
 
 
-class ResellerTierManager(NamedRecordManagerBase[ResellerTier]):
+class ResellerTierManager(
+    RecordManagerBase[ResellerTier],
+    NamedRecordManagerMixin[ResellerTier],
+):
     env_name = "openstack.reseller.tier"
     record_class = ResellerTier
