+++ openstack_odooclient/managers/support_subscription.py
@@ -18,8 +18,9 @@ from __future__ import annotations
 from datetime import date
 from typing import Annotated, Literal
 
-from ..base.record import ModelRef, RecordBase
-from ..base.record_manager import RecordManagerBase
+from ..base.record.base import RecordBase
+from ..base.record.types import ModelRef
+from ..base.record_manager.base import RecordManagerBase
 
 
 class SupportSubscription(RecordBase["SupportSubscriptionManager"]):
