+++ openstack_odooclient/managers/support_subscription_type.py
@@ -17,16 +17,18 @@ from __future__ import annotations
 
 from typing import Annotated, Literal
 
-from ..base.record import ModelRef, RecordBase
-from ..base.record_manager_named import NamedRecordManagerBase
+from ..base.record.base import RecordBase
+from ..base.record.types import ModelRef
+from ..base.record_manager.base import RecordManagerBase
+from ..mixins.named_record import NamedRecordManagerMixin, NamedRecordMixin
 
 
-class SupportSubscriptionType(RecordBase["SupportSubscriptionTypeManager"]):
+class SupportSubscriptionType(
+    RecordBase["SupportSubscriptionTypeManager"],
+    NamedRecordMixin["SupportSubscriptionTypeManager"],
+):
     billing_type: Literal["paid", "complimentary"]
     """The type of support subscription."""
-
-    name: str
-    """The name of the support subscription type."""
 
     product_id: Annotated[int, ModelRef("product", Product)]
     """The ID for the product to use to invoice
@@ -73,7 +75,8 @@ class SupportSubscriptionType(RecordBase["SupportSubscriptionTypeManager"]):
 
 
 class SupportSubscriptionTypeManager(
-    NamedRecordManagerBase[SupportSubscriptionType],
+    RecordManagerBase[SupportSubscriptionType],
+    NamedRecordManagerMixin[SupportSubscriptionType],
 ):
     env_name = "openstack.support_subscription.type"
     record_class = SupportSubscriptionType
