+++ openstack_odooclient/managers/term_discount.py
@@ -20,8 +20,9 @@ from typing import Annotated
 
 from typing_extensions import Self
 
-from ..base.record import ModelRef, RecordBase
-from ..base.record_manager import RecordManagerBase
+from ..base.record.base import RecordBase
+from ..base.record.types import ModelRef
+from ..base.record_manager.base import RecordManagerBase
 
 
 class TermDiscount(RecordBase["TermDiscountManager"]):
