+++ openstack_odooclient/managers/uom_category.py
@@ -17,8 +17,8 @@ from __future__ import annotations
 
 from typing import Literal
 
-from ..base.record import RecordBase
-from ..base.record_manager import RecordManagerBase
+from ..base.record.base import RecordBase
+from ..base.record_manager.base import RecordManagerBase
 
 
 class UomCategory(RecordBase["UomCategoryManager"]):
