+++ openstack_odooclient/mixins/named_record.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Catalyst Cloud Limited
+# Copyright (C) 2025 Catalyst Cloud Limited
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,38 +15,41 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Literal, overload
+from typing import TYPE_CHECKING, Any, Generic, Literal, overload
 
-from .record_manager_with_unique_field import (
-    Record,
-    RecordManagerWithUniqueFieldBase,
-)
+from ..base.record.base import RM, RecordProtocol
+from ..base.record_manager.base import R, RecordManagerProtocol
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
 
 
-class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
-    """A record manager base class for record types with a code field.
+class NamedRecordMixin(RecordProtocol[RM], Generic[RM]):
+    """A record mixin for record types with a ``name`` field.
 
-    This code field is reasonably expected to be unique, which allows
-    methods for getting records by code to be defined.
-
-    The record class should be type hinted with the field to use as the code,
-    just like any other field.
-    Configure the name of the code field on the manager class by defining
-    the ``code_field`` attribute (set to ``code`` by default).
+    Include this mixin to add the ``name`` field, with ``str`` type,
+    to your custom record class. This allows the additional methods
+    provided by ``NamedRecordManagerMixin`` to be used.
     """
 
-    code_field: str = "code"
-    """The field code to use when querying by code in
-    the ``get_by_code`` method.
+    name: str
+    """The unique name of the record."""
+
+
+class NamedRecordManagerMixin(RecordManagerProtocol[R], Generic[R]):
+    """A record manager mixin for record types with a ``name`` field.
+
+    The ``name`` field is expected to be unique, which allows
+    methods for getting records by name to be defined.
+
+    The record class must either include the ``NamedRecordMixin`` mixin,
+    or have its own ``name`` field defined with the ``str`` type.
     """
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[True],
@@ -55,9 +58,9 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> int | None: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[True],
@@ -66,9 +69,9 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> int | None: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[True],
@@ -77,9 +80,9 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> int: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[True],
@@ -88,9 +91,9 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> int: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[False] = ...,
@@ -99,9 +102,9 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> dict[str, Any] | None: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[False] = ...,
@@ -110,47 +113,47 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
     ) -> dict[str, Any]: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[False] = ...,
         as_dict: Literal[False] = ...,
         optional: Literal[True],
-    ) -> Record | None: ...
+    ) -> R | None: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: Literal[False] = ...,
         as_dict: Literal[False] = ...,
         optional: Literal[False] = ...,
-    ) -> Record: ...
+    ) -> R: ...
 
     @overload
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         *,
         fields: Iterable[str] | None = ...,
         as_id: bool = ...,
         as_dict: bool = ...,
         optional: bool = ...,
-    ) -> Record | int | dict[str, Any] | None: ...
+    ) -> R | int | dict[str, Any] | None: ...
 
-    def get_by_code(
+    def get_by_name(
         self,
-        code: str,
+        name: str,
         fields: Iterable[str] | None = None,
         as_id: bool = False,
         as_dict: bool = False,
         optional: bool = False,
-    ) -> Record | int | dict[str, Any] | None:
-        """Query a unique record by code.
+    ) -> R | int | dict[str, Any] | None:
+        """Query a unique record by name.
 
         A number of parameters are available to configure the return type,
         and what happens when a result is not found.
@@ -166,10 +169,10 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
         a ``dict`` object, instead of a record object.
 
         When ``optional`` is ``True``, ``None`` is returned if a record
-        with the given code does not exist, instead of raising an error.
+        with the given name does not exist, instead of raising an error.
 
-        :param code: The record code
-        :type code: str
+        :param name: The record name
+        :type name: str
         :param as_id: Return a record ID, defaults to False
         :type as_id: bool, optional
         :param fields: Fields to select, defaults to ``None`` (select all)
@@ -178,14 +181,14 @@ class CodedRecordManagerBase(RecordManagerWithUniqueFieldBase[Record, str]):
         :type as_dict: bool, optional
         :param optional: Return ``None`` if not found, defaults to False
         :type optional: bool, optional
-        :raises MultipleRecordsFoundError: Multiple records with the same code
-        :raises RecordNotFoundError: Record with the given code not found
+        :raises MultipleRecordsFoundError: Multiple records with the same name
+        :raises RecordNotFoundError: Record with the given name not found
         :return: Query result (or ``None`` if record not found and optional)
-        :rtype: Record | int | dict[str, Any] | None
+        :rtype: R | int | dict[str, Any] | None
         """
-        return self._get_by_unique_field(
-            field=self.code_field,
-            value=code,
+        return self.get_by_unique_field(
+            field="name",
+            value=name,
             fields=fields,
             as_id=as_id,
             as_dict=as_dict,
